Move referencedUndeclaredNames into common; dedupe slow second copy
init_region.mjs had a second copy of referencedUndeclaredNames that
still used the slow babel-traverse + clone-into-synthetic-File pattern;
the boundary.mjs version had been replaced earlier with a hand-rolled
recursive walker. Moved the optimized walker into
common/program_analysis.mjs and made both call sites import it.

Profile-driven attempts at the rest of the proposed Tier 1 wins:

- Removing programPath.scope.crawl() in analyzeRuntimeBoundaryAst was
  actually 30% slower: every per-visitor path.scope.getBinding() call
  triggers an ad-hoc partial crawl, which adds up to more work than
  one upfront crawl over the program. Reverted.
- The array-spread allocations in decl_graph.mjs buildStagedShellBatchPlan
  are real but each batch plan touches small (~10s of items) arrays.
  Replacing them in place doesn't move the needle on the benchmark
  wall measurably; the real cost in the planner is graph construction,
  not the spreads.

Per-chunk timings on the Excalidraw entry chunk:

  lower      454ms -> 327ms   (init_region picks up optimized walker)
  total wall ~11.1s -> ~10.9s  (within run-to-run noise)

Honest assessment: the meaningful next step is the Tier 2 rewrite
(custom AST walker replacing babel-traverse for the whole boundary
analysis). That's the path to <5s wall.

BAZEL_TEST_INVOCATIONS=local:267b200f-098c-4bcb-9a29-fbaf4065dc3e
