+++ devinfra/js/debundle/analysis/boundary.mjs
@@ -6,6 +6,7 @@ import * as t from "@babel/types";
 import { DEFAULT_PARSER_OPTIONS, writeJsonFile } from "../common/parser_options.mjs";
 import { getArtifactManifestChunks, getChunkEntryFile, getChunkEntryPath, requirePipelineArtifact } from "../common/artifact.mjs";
 import { formatDurationSince, logProgress, prepareOutputDir, relativeWorkspacePath, resolveWorkspacePath } from "../common/io.mjs";
+import { referencedUndeclaredNames } from "../common/program_analysis.mjs";
 
 const traverse = traverseModule.default ?? traverseModule;
 
@@ -1220,81 +1221,6 @@ function effectToReason(effect) {
     return "contains_top_level_await";
   }
   return effect;
-}
-
-// Names referenced in `node` from positions that are NOT inside a nested
-// function/method body. Equivalent semantics to a babel-traverse pass with
-// a ReferencedIdentifier visitor that filters out names with a binding in
-// scope; in our use site `node` is a VariableDeclarator init expression,
-// so the only "binding in scope" possibilities are inside nested functions
-// (which we skip anyway). Hand-rolled recursion is ~10x cheaper than
-// wrapping in a synthetic File and running traverse with scope.
-function referencedUndeclaredNames(node) {
-  if (!node) return [];
-  const names = new Set();
-  collectReferencedIdentifierNames(node, names);
-  return [...names];
-}
-
-const FUNCTION_TYPES = new Set([
-  "FunctionExpression",
-  "FunctionDeclaration",
-  "ArrowFunctionExpression",
-  "ObjectMethod",
-  "ClassMethod",
-  "ClassPrivateMethod",
-]);
-
-const NON_AST_KEYS = new Set([
-  "type",
-  "start",
-  "end",
-  "loc",
-  "extra",
-  "leadingComments",
-  "trailingComments",
-  "innerComments",
-  "comments",
-  "tokens",
-  "errors",
-]);
-
-function collectReferencedIdentifierNames(node, names) {
-  if (!node || typeof node !== "object") return;
-  if (Array.isArray(node)) {
-    for (const child of node) collectReferencedIdentifierNames(child, names);
-    return;
-  }
-  if (FUNCTION_TYPES.has(node.type)) return;
-  switch (node.type) {
-    case "Identifier":
-      names.add(node.name);
-      return;
-    case "MemberExpression":
-    case "OptionalMemberExpression":
-      collectReferencedIdentifierNames(node.object, names);
-      if (node.computed) collectReferencedIdentifierNames(node.property, names);
-      return;
-    case "ObjectProperty":
-      if (node.computed) collectReferencedIdentifierNames(node.key, names);
-      collectReferencedIdentifierNames(node.value, names);
-      return;
-    case "VariableDeclarator":
-      // node.id is the declaration target (not a reference)
-      collectReferencedIdentifierNames(node.init, names);
-      return;
-    case "ImportSpecifier":
-    case "ImportDefaultSpecifier":
-    case "ImportNamespaceSpecifier":
-    case "ExportSpecifier":
-      return;
-    default: {
-      for (const key of Object.keys(node)) {
-        if (NON_AST_KEYS.has(key)) continue;
-        collectReferencedIdentifierNames(node[key], names);
-      }
-    }
-  }
 }
 
 function sortedAccessNames(accessMap) {
